Split functionality in packages
